pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neko"
version = "2.0.0"
dependencies = [
    "rich",
    "requests",
    "dnspython",
    "paramiko",
    "python-whois",
    "shodan"
]

